+++ net.h
@@ -9,7 +9,7 @@
 struct net {
     void *nic;
     uint8_t addr[ETHER_ADDR_LEN];
-    uint32_t (*send)(struct net *inet, void *p_packet, size_t packet_size);
+    uint32_t (*send)(struct net *inet, void *p_packet, uint32_t packet_size);
 };
